Move some of the utilities that are shared between the different plotting scripts into the utilities directory
