+++ src/postgkyl/output/__init__.py
@@ -2,8 +2,4 @@
 from .plot import plot
 from .plotly import plotly_animate, plotly
 from .pyvista import pyvista
-from .downsample import downsample
-from .nodal_to_cell_centered_grid import nodal_to_cell_centered_grid
-from .axis_and_grid_prep import axis_and_grid_prep
-from .load_plot_data import load_plot_data
 from .plot import pgkyl_colorbar
