+++ src/postgkyl/utils/__init__.py
@@ -2,3 +2,7 @@ from .input_parser import input_parser
 from .load_style import load_style
 from .verb_print import verb_print
 from .set_frame import set_frame
+from .downsample import downsample
+from .nodal_to_cell_centered_grid import nodal_to_cell_centered_grid
+from .axis_and_grid_prep import axis_and_grid_prep
+from .load_plot_data import load_plot_data
